Sample project for Arduinos' I2C bus
In this project, the master is connected to 2 slaves via I2C bus. The
master will count from 0 to 100, then, for each count, it will send the
even number to slave1 to print on serial terminal and odd number to
slave2 to print on a different serial terminal. At the end of the
process. The slaves will send a finish message back to master to print
it on serial terminal.